# lib/api-zod/src/generated/types/createServiceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trançify SaaS multi-tenant scheduling API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateServiceRequest {
  name: string;
  description?: string;
  durationHours: number;
  priceSmall: number;
  priceLarge: number;
  sizeDependent: boolean;
}