# lib/api-zod/src/generated/types/tenant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trançify SaaS multi-tenant scheduling API
 * OpenAPI spec version: 0.1.0
 */
import type { TenantPlan } from "./tenantPlan";
import type { TenantStatus } from "./tenantStatus";

export interface Tenant {
  id: string;
  slug: string;
  name: string;
  email: string;
  whatsapp?: string | null;
  logoUrl?: string | null;
  primaryColor?: string | null;
  secondaryColor?: string | null;
  status: TenantStatus;
  plan: TenantPlan;
  createdAt: Date;
}