# lib/api-zod/src/generated/types/updateTenantRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trançify SaaS multi-tenant scheduling API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTenantRequest {
  name?: string;
  /**
   * @minLength 3
   * @maxLength 50
   * @pattern ^[a-z0-9]+(?:-[a-z0-9]+)*$
   */
  slug?: string;
  whatsapp?: string;
  logoUrl?: string;
  primaryColor?: string;
  secondaryColor?: string;
}